Create a plan for proxying builder-api
This commit adds a plan and configuration for standing up nginx in
front of the monolithic depot. It uses a slightly modified (as in,
it's a template) nginx.conf from the monolithic system, and should be
added to the director configuration used to run the monolith. It will
be refactored for discovery and all the goods at a future date.

Signed-off-by: jtimberman <joshua@chef.io>

Pull request: #898
Approved by: smith